Updated tests to use new bundling infra, allowing us to remove private exports.
